Connect to mongoDB atlas

+++ app.js
@@ -23,16 +23,25 @@ app.use(bodyParser.urlencoded({
 }));
 
 // MongoDB URL
-const url = "mongodb://localhost:27017"
+// const url = "mongodb://localhost:27017"
 
 // DB Name
 const dbName = "todolistDB"
 
-mongoose.connect(url + "/" + dbName, {
-  useUnifiedTopology: true,
-  useNewUrlParser: true,
-  useFindAndModify: false,
-});
+
+/* when connect to MongoDB atlas, if the connection string contains any special
+characters, will have to replace it with encoded character, in this case, 
+replacing "#" with "%23", for more encoding detail, see https://www.urlencoder.org/ */
+
+mongoose.connect(
+  "mongodb+srv://admin-darren:Mongodb382%23@cluster0-gmncq.mongodb.net/" +
+    dbName,
+  {
+    useUnifiedTopology: true,
+    useNewUrlParser: true,
+    useFindAndModify: false,
+  }
+);
 
 const itemSchema = new mongoose.Schema({
   name: String
@@ -75,15 +84,35 @@ app.get("/", (req, res) => {
   /* render the file name "list.ejs" under "view/" directory.a
   use ejs file to replace html file, but ejs file have same functionality of html files*/
 
-  Item.find({}, (err, listItem) => {
+  /* model "Item" is mongoose representation of MongoDB collection "items".
+    "model.find({})" will find all the items in given collection.
+  */
+
+  /* need to find a way to render navbar page so we have a drop down list of all 
+  list collection in database. */
+  // console.log(List.find({}, (err, doc) => {
+  //   if (!err) {
+  //     return doc
+  //   }
+  // }))
+
+
+
+
+  Item.find({}, (err, foundItems) => {
     if (err) {
 
       console.log(err);
 
-    } else if (listItem.length === 0) {
+      /* "Item.find({}, (err, foundItem)" will return a list of items in the collections
+       and log them in "foundItems" variable */
+    } else if (foundItems.length === 0) {
 
-      /* check if listItem is empty, if it's empty, it will insert the some default
-      items in the listItems, so it can show up on the /home page. 
+
+      // console.log(foundItem)
+
+      /* check if foundItem is empty, if it's empty, it will insert the some default
+      items in the foundItems, so it can show up on the root route. 
       This also prevents the code from adding the defaultItems to DB everytime code restarted*/
       Item.insertMany(defaultItems, (err) => {
         if (err) {
@@ -92,14 +121,14 @@ app.get("/", (req, res) => {
           console.log("default Items has been saved to DB.");
         }
       });
-      /* re-enter /home page to have the default items show up on screen. */
-      res.redirect("/");
+      /* re-enter root route page to have the default items show up on screen. */
+        res.redirect("/");
     } else {
       /* if default list of item is saved to DB, it will not save them again, 
-      instead, it will show the defaultList items on the scrren */
+      instead, it will show the defaultList items on the screen */
       res.render("list", {
         listTitle: "Today",
-        newListItems: listItem
+        newListItems: foundItems
       });
     }
   })
@@ -110,8 +139,8 @@ app.get("/", (req, res) => {
 
 app.get("/:listName", (req, res) => {
   /* Dynamic route name using express parameter feature */
-  console.log(req.params.listName);
-  const listName = req.params.listName;
+  // console.log(req.params.listName);
+  const listName = _.capitalize(req.params.listName);
 
   List.findOne({
     name: listName
@@ -123,7 +152,7 @@ app.get("/:listName", (req, res) => {
           items: defaultItems
         });
         list.save();
-        res.redirect("/"+listName)
+        res.redirect("/" + listName)
       } else {
         res.render("list", {
           listTitle: foundList.name,
@@ -155,7 +184,7 @@ app.get("/about", (req, res) => {
 
 app.post("/", (req, res) => {
 
-  console.log(req.body);
+  // console.log(req.body);
 
   const itemName = req.body.newItem;
   const listName = req.body.list;
@@ -168,21 +197,23 @@ app.post("/", (req, res) => {
     item.save();
     res.redirect("/");
   } else {
-    List.findOne({ name: listName }, (err, foundList) => {
+    List.findOne({
+      name: listName
+    }, (err, foundList) => {
       foundList.items.push(item);
       foundList.save();
-      res.redirect("/"+ listName)
+      res.redirect("/" + listName)
     })
   }
 
   /* ".save()" method will work as native MongoDB method ".insertOne()", 
   the purpose of this ".save()" method is as its name, save the document \
   to collection.*/
-  
+
 
   /* alway use res.redirect("/") to go back to home page in order to refresh
   load of DB documents to show up */
-  
+
   /* COMMENT out ordinary add item, since this is no longer a static array 
   declared in the beginning of this code. */
   // if (req.body.list === "Work") {
@@ -201,18 +232,44 @@ app.post("/delete", (req, res) => {
   if item in collection can be deleted. */
 
   const checkItemId = req.body.checkedBox;
-  console.log(req.body.checkedBox);
+  const listName = req.body.listName
 
-  /* Here can use deleteOne method or to use findByIdAndRemove method */
+  /* check if it's default list you are working on. */
+  if (listName === "Today") {
+    /* Here can use deleteOne method or to use findByIdAndRemove method */
 
-  Item.findByIdAndRemove(checkItemId, (err) => {
-    if (err) {
-      console.log(err);
-    } else {
-      console.log("Checked Item Successfully Deleted");
-      res.redirect("/");
-    }
-  });
+    Item.findByIdAndRemove(checkItemId, (err) => {
+      if (err) {
+        console.log(err);
+      } else {
+        console.log("Checked Item Successfully Deleted");
+        res.redirect("/");
+      }
+    });
+  } else {
+    /* a easier way is to use .findOneAndUpdate and $pull combination in this case */
+    List.findOneAndUpdate({
+        name: listName
+      },
+      /* $pull from "items" array from "list" collection, in which "_id" of an 
+      item is "ItemId" we got from above.  $pull will remove item from array. 
+      Check MongoDB and Mongoose doc for more detail*/
+
+      {
+        $pull: {
+          items: {
+            _id: checkItemId
+          }
+        }
+      },
+      (err, foundList) => {
+        if (!err) {
+          res.redirect("/" + listName)
+        }
+      });
+  }
+
+
 
   /* either of these should work, but .findByIdAndRemove would be more straightforward. */
